Add user input to file save and load
Adds new method to offer the user a chance to enter a file name to load on start
up or when selecting 'load students' on menu. Same method is used for selecting
file to save to.

A class variable keeps track of the loaded file and lets the user save it to
this by hitting return. Variable set to default 'students.csv' on loading
application.