Fix load file method
`Load_file` method (previously `load_students`) previously failed when the file
entered didn't exist, as the if statement was located after the method running.

This now works, and there is an additional 'loaded blank file' message if there
is no file to load.

Issue: there is now no option to load a blank file from the outset. Only loads if no
file is entered and no 'students.csv' is found.