Update to ask user to choose load method
Adds `welcome` method which runs if no argument has been provided in the command
line and  asks user to select what to load when opening app:

1. Load blank directory
2. Load students.csv
3. Load another file

Further refactoring required. Also need to fix method currently running when
user chooses to load a new file, as it starts a new session.